sandwich-shop-manager/src/main/java/it/unipd/tos/business/Bill.java
////////////////////////////////////////////////////////////////////
// [Nicolo] [Frison] [1147682]
////////////////////////////////////////////////////////////////////
package it.unipd.tos.business;

import java.util.List;

import it.unipd.tos.business.exception.TakeAwayBillException;
import it.unipd.tos.model.MenuItem;

public class Bill implements TakeAwayBill {
    public double getOrderPrice(List<MenuItem> itemsOrdered) throws TakeAwayBillException {
        double total = 0;
        
        for( MenuItem i : itemsOrdered ) {
            total += i.getPrice();
        }
        
        return total;
    }
}
